move userProvider from folder context
